requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
backcall==0.2.0
beautifulsoup4==4.11.1
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-bootstrap4==22.1
django-braces==1.15.0
django-debug-toolbar==3.5.0
executing==0.8.3
misaka==2.1.1
pickleshare==0.7.5
Pillow==9.2.0
psycopg2 @ file:///C:/ci_310/psycopg2_1642085998310/work
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
pywin32==303
pyzmq==22.3.0
soupsieve==2.3.2.post1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tornado==6.1
traitlets==5.1.1
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
wcwidth==0.2.5
wincertstore==0.2
